InterpreterPart1.js is my representation of Ruslan's blog in js.
Following along with Ruslan's blog converting to js. I have solutions to the questions he asks his reader at the end of his blogs. They can be found by ctrl-F and searching for --Q.
Interpreter2.js is incomplete. I added the interpreter.js file to my html to run in the browser instead of the command line. Just a matter of preference. You can uncomment the code at the top of interpreter files and run the code in the command line.